iswap/dashboard/sendemail.py
from iswap import app
from flask_mail import Mail, Message

app.config['MAIL_SERVER']='smtp.gmail.com'
app.config['MAIL_PORT'] = 465
app.config['MAIL_USERNAME'] = 'oyarojared278@gmail.com'
app.config['MAIL_PASSWORD'] = 'miami2015'
app.config['MAIL_USE_TLS'] = False
app.config['MAIL_USE_SSL'] = True
mail = Mail(app)

